api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
anthropic>=0.40.0
httpx>=0.27.0
feedparser>=6.0.11
pyyaml>=6.0.1
apscheduler>=3.10.0,<4.0
